cross state is not yet what i wanted :s
