+++ account/server/src/auth/command.rs
@@ -2,7 +2,7 @@ use crate::auth::{get_key, JWT_ISSUER, JWT_SECRET};
 use account_state::error::AccountError;
 
 use jsonwebtokens as jwt;
-use jsonwebtokens::{encode, AlgorithmID};
+use jsonwebtokens::{AlgorithmID, encode};
 use jwt::Algorithm;
 use rocket::serde::json::Json;
 use rocket::State;
@@ -13,7 +13,8 @@ use crate::auth::jwt_guard::JwtToken;
 use crate::MariadDb;
 use account_shared::{AccountCommand, CreateAccount, Login};
 use account_state::state::AccountState;
-use state_repository::{ModelKey, StateRepository};
+use state_repository::StateRepository;
+use state_repository::model_key::ModelKey;
 
 #[post("/", format = "json", data = "<command>")]
 pub async fn handle_anonymous(
