+++ lib/state-repository/tests/cross_state/mod.rs
@@ -1,0 +1,5 @@
+pub mod build;
+pub mod flow;
+pub mod gold;
+
+use serde::{Deserialize, Serialize};
