--- lib/workflow/Cargo.toml
@@ -1,17 +1,0 @@
-[package]
-name = "workflow"
-version = "0.1.0"
-edition = "2021"
-
-# See more keys and their definitions at https://doc.rust-lang.org/cargo/reference/manifest.html
-
-[dependencies]
-state = { path = "../state" }
-anyhow = "1.0"
-eventstore = "2.1"
-redis= { version = "0.22", features = ["tokio-native-tls-comp"] }
-serde = { version = "1.0", features = ["derive"] }
-serde_json = "1"
-uuid = {version = "0.8", features = ["v4", "serde"]}
-tokio = "1.21"
-state-repository = { path = "../state-repository" }
