--- lib/workflow/tests/multiple/mod.rs
@@ -1,12 +1,0 @@
-pub mod build;
-pub mod flow;
-pub mod gold;
-pub mod worker;
-
-use serde::{Deserialize, Serialize};
-
-#[derive(Deserialize, Serialize, Default, Debug, Copy, Clone, Eq, PartialEq)]
-pub struct Cost {
-    pub gold: u32,
-    pub worker: u32,
-}
